Cannot use an abstract base class with hibernate like this ...
