+++ aclslib/src/main/java/au/edu/uq/cmm/aclslib/server/Configuration.java
@@ -26,5 +26,13 @@ public interface Configuration {
      String getBaseFileUrl();
      
      String getCaptureDirectory();
+     
+     String getFeedId();
+     
+     String getFeedTitle();
+     
+     String getFeedAuthor();
+     
+     String getFeedAuthorEmail();
 
 }
